Move Replayer.tsx
